requirements.txt
aiogram==3.5.0
python-dotenv==1.0.1
openai==1.30.1
supabase==1.2.0
redis==5.0.1
aioredis==2.0.1
cryptography==42.0.7
pydub>=0.25.1
ffmpeg-python